Case 784 - Upgraded to LevelDB 1.18

+++ db/autocompact_test.cc
@@ -88,10 +88,10 @@ void AutoCompactTest::DoReads(int n) {
     delete iter;
     // Wait a little bit to allow any triggered compactions to complete.
     Env::Default()->SleepForMicroseconds(1000000);
-    int64_t size = Size(Key(0), Key(n));
+    uint64_t size = Size(Key(0), Key(n));
     fprintf(stderr, "iter %3d => %7.3f MB [other %7.3f MB]\n",
             read+1, size/1048576.0, Size(Key(n), Key(kCount))/1048576.0);
-    if (size <= initial_size/4) { // a factor of 10 is too great
+    if (size <= static_cast<uint64_t>(initial_size/4)) { // 10 is too much, put it to 4
       break;
     }
   }
